trying to fix potential performance bug in syntactic_unify:
in most cases, recursion under fun (and let) did not actually
take place, and instead, the default case was invoked (which
works, but calls unify, which might be too expensive).

Got stuck trying to fix it on the error "Variable f should be
bound to a term but is bound to a constr_under_binders.",
where "f" is the rhs of a Gallina match clause, not an
argument to a tactic.

Also copied tests from
https://github.com/andres-erbsen/coq-experiments/blob/master/experiments/ltac1core.v